fix: feed image placeholder (#256)
* fix: update fcm token field name

* fix: logout api matching

* fix: show loading image indicator using cached_network_image

+++ lib/app/modules/auth/data/repositories/remote_auth_repository.dart
@@ -42,7 +42,9 @@ class RemoteAuthRepository implements AuthRepository {
 
   @override
   Future<void> logout() async {
-    await _api.logout();
+    final accessToken = await _tokenRepository.token.first;
+    if (accessToken == null) return;
+    await _api.logout(accessToken);
     await _tokenRepository.deleteToken();
     _userController.add(null);
   }
